third_party/blink/web_tests/svg/custom/use-cycle-with-attr-eventhandler-expected.txt

CONSOLE MESSAGE: line 4: foo is called
